Revert Async changes to SNIPacket to fix deadlock issues
